+++ Sources/KeyboardKit/Layout/Providers/iPhoneKeyboardLayoutProvider.swift
@@ -28,7 +28,8 @@ open class iPhoneKeyboardLayoutProvider: SystemKeyboardLayoutProvider {
      */
     open override func actions(
         for inputs: InputSetRows,
-        context: KeyboardContext) -> KeyboardActionRows {
+        context: KeyboardContext
+    ) -> KeyboardActionRows {
         let actions = super.actions(for: inputs, context: context)
         guard isExpectedPhoneInputActions(actions) else { return actions }
         let upper = actions[0]
@@ -50,10 +51,11 @@ open class iPhoneKeyboardLayoutProvider: SystemKeyboardLayoutProvider {
         for action: KeyboardAction,
         row: Int,
         index: Int,
-        context: KeyboardContext) -> KeyboardLayoutItemWidth {
+        context: KeyboardContext
+    ) -> KeyboardLayoutItemWidth {
         if action.isPrimaryAction { return bottomPrimaryButtonWidth(for: context) }
         switch action {
-        case dictationReplacement: return bottomSystemButtonWidth(for: context)
+        case context.keyboardDictationReplacement: return bottomSystemButtonWidth(for: context)
         case .character:
             if context.isAlphabetic(.greek) { return .percentage(0.1) }
             return isLastNumericInputRow(row, for: context) ? lastSymbolicInputWidth(for: context) : .input
@@ -73,13 +75,16 @@ open class iPhoneKeyboardLayoutProvider: SystemKeyboardLayoutProvider {
     /**
      Get the actions of the bottommost space key row.
      */
-    open func bottomActions(for context: KeyboardContext) -> KeyboardActions {
+    open func bottomActions(
+        for context: KeyboardContext
+    ) -> KeyboardActions {
         var result = KeyboardActions()
         let needsInputSwitch = context.needsInputModeSwitchKey
         let needsDictation = context.needsInputModeSwitchKey
         if let action = keyboardSwitchActionForBottomRow(for: context) { result.append(action) }
         if needsInputSwitch { result.append(.nextKeyboard) }
         if !needsInputSwitch { result.append(.keyboardType(.emojis)) }
+        let dictationReplacement = context.keyboardDictationReplacement
         if isPortrait(context), needsDictation, let action = dictationReplacement { result.append(action) }
         result.append(.space)
         if context.isAlphabetic(.persian) { result.append(.character(.zeroWidthSpace)) }
@@ -91,7 +96,10 @@ open class iPhoneKeyboardLayoutProvider: SystemKeyboardLayoutProvider {
     /**
      Get leading actions to add to the lower inputs row.
      */
-    open func lowerLeadingActions(for actions: KeyboardActionRows, context: KeyboardContext) -> KeyboardActions {
+    open func lowerLeadingActions(
+        for actions: KeyboardActionRows,
+        context: KeyboardContext
+    ) -> KeyboardActions {
         guard isExpectedPhoneInputActions(actions) else { return [] }
         let margin = leadingMarginAction(for: actions[2])
         if context.isAlphabetic(.hebrew) { return [margin] }
@@ -111,7 +119,10 @@ open class iPhoneKeyboardLayoutProvider: SystemKeyboardLayoutProvider {
     /**
      Get trailing actions to add to the lower inputs row.
      */
-    open func lowerTrailingActions(for actions: KeyboardActionRows, context: KeyboardContext) -> KeyboardActions {
+    open func lowerTrailingActions(
+        for actions: KeyboardActionRows,
+        context: KeyboardContext
+    ) -> KeyboardActions {
         guard isExpectedPhoneInputActions(actions) else { return [] }
         let margin = trailingMarginAction(for: actions[2])
         if context.isAlphabetic(.hebrew) { return [margin] }
@@ -129,7 +140,10 @@ open class iPhoneKeyboardLayoutProvider: SystemKeyboardLayoutProvider {
     /**
      Get leading actions to add to the middle inputs row.
      */
-    open func middleLeadingActions(for actions: KeyboardActionRows, context: KeyboardContext) -> KeyboardActions {
+    open func middleLeadingActions(
+        for actions: KeyboardActionRows,
+        context: KeyboardContext
+    ) -> KeyboardActions {
         guard shouldAddMiddleMarginActions(for: actions, context: context) else { return [] }
         let margin = leadingMarginAction(for: actions[1])
         return [margin]
@@ -138,7 +152,10 @@ open class iPhoneKeyboardLayoutProvider: SystemKeyboardLayoutProvider {
     /**
      Get trailing actions to add to the middle inputs row.
      */
-    open func middleTrailingActions(for actions: KeyboardActionRows, context: KeyboardContext) -> KeyboardActions {
+    open func middleTrailingActions(
+        for actions: KeyboardActionRows,
+        context: KeyboardContext
+    ) -> KeyboardActions {
         guard shouldAddMiddleMarginActions(for: actions, context: context) else { return [] }
         let margin = trailingMarginAction(for: actions[1])
         return [margin]
@@ -147,7 +164,10 @@ open class iPhoneKeyboardLayoutProvider: SystemKeyboardLayoutProvider {
     /**
      Get leading actions to add to the upper inputs row.
      */
-    open func upperLeadingActions(for actions: KeyboardActionRows, context: KeyboardContext) -> KeyboardActions {
+    open func upperLeadingActions(
+        for actions: KeyboardActionRows,
+        context: KeyboardContext
+    ) -> KeyboardActions {
         let margin = leadingMarginAction(for: actions[0])
         if context.isAlphabetic(.hebrew) { return [margin] }
         guard shouldAddUpperMarginActions(for: actions, context: context) else { return [] }
@@ -157,7 +177,10 @@ open class iPhoneKeyboardLayoutProvider: SystemKeyboardLayoutProvider {
     /**
      Get trailing actions to add to the upper inputs row.
      */
-    open func upperTrailingActions(for actions: KeyboardActionRows, context: KeyboardContext) -> KeyboardActions {
+    open func upperTrailingActions(
+        for actions: KeyboardActionRows,
+        context: KeyboardContext
+    ) -> KeyboardActions {
         let margin = trailingMarginAction(for: actions[0])
         if context.isAlphabetic(.hebrew) { return [margin, .backspace] }
         guard shouldAddUpperMarginActions(for: actions, context: context) else { return [] }